requirements.txt
asgiref==3.6.0
distlib==0.3.6
Django==4.1.7
django-filter==23.2
djangorestframework==3.14.0
filelock==3.9.0
gunicorn==20.1.0
Markdown==3.4.3
platformdirs==3.0.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.3
tzdata==2022.7
virtualenv==20.19.0
virtualenvwrapper-win==1.2.7
waitress==2.1.2